Add stres tests to release process (#348)
* Add stres tests to release process

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

* wip

+++ scripts/testing/stress-mem.sh
@@ -22,19 +22,34 @@ do echo -e "\n---Waiting for BTP Operator to be ready and reconciled"; sleep 5; 
 echo -e "\n---Creating ${N} service bindings and instances"
 
 for ((i=1; i <= N ; i++))
-do
-    SI_NAME=auditlog-management-si-$i
-    SB_NAME=auditlog-management-sb-$i
+  do
+      SI_NAME=auditlog-management-si-$i
+      SB_NAME=auditlog-management-sb-$i
 
-    export SI_NAME
-    export SB_NAME
+      export SI_NAME
+      export SB_NAME
 
-    envsubst <${YAML_DIR}/e2e-test-service-instance.yaml | kubectl apply -f - >/dev/null
-    envsubst <${YAML_DIR}/e2e-test-service-binding.yaml | kubectl apply -f - >/dev/null
+      envsubst <${YAML_DIR}/e2e-test-service-instance.yaml | kubectl apply -f - >/dev/null
+      envsubst <${YAML_DIR}/e2e-test-service-binding.yaml | kubectl apply -f - >/dev/null
 done
 
 echo -e "\n---${N} service bindings and instances created - let them be for a while... ${LIFE_SPAN}s"
 sleep ${LIFE_SPAN}
+
+restarts=$(kubectl get po -n kyma-system -l app.kubernetes.io/component=btp-manager.kyma-project.io -o 'jsonpath={..items[0].status.containerStatuses[0].restartCount}')
+if [ "${restarts}" != '0' ]
+then
+  echo "BTP manager was restarted $restarts times"
+  exit 1
+fi
+
+restarts=$(kubectl get po -n kyma-system -l app.kubernetes.io/name=sap-btp-operator -o 'jsonpath={..items[0].status.containerStatuses[?(@.name=="manager")].restartCount}')
+if [ "${restarts}" != '0' ]
+then
+  echo "BTP operator was restarted $restarts times"
+  exit 1
+fi
+
 
 echo -e "\n---Deleting e2e-test-btpoperator"
 kubectl delete btpoperators/e2e-test-btpoperator
